refactor: 마크다운 컨텐츠 컴포넌트 분리
