+++ src/app/(default)/posts/[slug]/_components/PostDetail.tsx
@@ -1,28 +1,6 @@
 import dayjs from "dayjs";
 import Link from "next/link";
-import Markdown, { type Components } from "react-markdown";
-
-const markdownComponents = {
-  h1: (props) => {
-    return <h1 className="mb-4 border-b text-3xl font-bold" {...props} />;
-  },
-  h2: (props) => {
-    return <h2 className="my-8 border-b pb-2 text-2xl font-bold" {...props} />;
-  },
-  h3: (props) => {
-    return <h3 className="my-8 text-xl font-bold" {...props} />;
-  },
-  img: (props) => {
-    // eslint-disable-next-line @next/next/no-img-element, jsx-a11y/alt-text
-    return <img className="mx-auto my-12" {...props} />;
-  },
-  p: (props) => {
-    return <p className="my-6" {...props} />;
-  },
-  a: (props) => {
-    return <a className="text-blue-500 hover:underline" {...props} />;
-  },
-} satisfies Components;
+import MarkdownContent from "@/components/posts/MarkdownContent";
 
 export default function PostDetail({ post }: { post: Post }) {
   return (
@@ -37,7 +15,7 @@ export default function PostDetail({ post }: { post: Post }) {
         </div>
       </div>
       <div>
-        <Markdown components={markdownComponents}>{post.content}</Markdown>
+        <MarkdownContent content={post.content} />
       </div>
     </div>
   );
